fix controller and add a test
